[added] a11y.restoreAll() method to clean up mutations to React
a11y patches the React.createElement method, and sometimes
patches lifecycle events of components. This adds a `restore`
method to `a11y` which cleans up after itself.

This is particularly useful for using React-a11y in JavaScript
test suites, where the patched React and/or components may be
rendered in future tests.